load sample data from external file
